Add solution for problem `A - Ezzat and Two Subsequences` Link: https://codeforces.com/contest/1557/problem/A Tags: brute force, math, sortings, *800 Ref: https://codeforces.com/contest/1557/submission/125406397

+++ codeforces/1557/A.cpp
@@ -8,7 +8,7 @@ const int MOD = 1000000007;
 void solve() {
 	int n;
 	cin >> n;
-	double a[n + 7];
+	long double a[n + 7];
 	for (int i = 0; i < n; i++) {
 		cin >> a[i];
 	}
